requirements.txt
amqp==5.1.1
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
kombu==5.3.2
numpy==1.26.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
outcome==1.2.0
Pillow==10.0.1
prompt-toolkit==3.0.39
PySocks==1.7.1
python-dateutil==2.8.2
pytube==15.0.0
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
watermark-remover-cli==0.0.1
wcwidth==0.2.8
wsproto==1.2.0
